Update file 102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 /**
+  * main - print first 50 fibonacci
   *
-  *
-  *
+  * Return: always 0
   */
 int main(void)
 {
-	int i, first = 0, second = 1, next = first + second;
+	long int i, first = 0, second = 1, next = first + second;
 
 	for (i = 0; i < 50; i++)
 	{
-		printf("%d", next);
+		printf("%ld", next);
 		if (i != 49)
 		{
 			printf(", ");
